WAITP-1278 Split the casing utils out from the general utils file
