+++ packages/types/src/utils/index.ts
@@ -5,4 +5,7 @@
  */
 
 export * from './vizTypes';
+export * from './casing';
+// Including utils/utils can break schema generation
+// Try including the specific file you want instead
 export * from './utils';
